Fixed Issues 68 and 106.
For Issue 68, all the agreed on changes in comments 5 and 10 were implemented.
Also added or modified a number of descriptive notes and design notes.

For Issue 106, the change was to add a description note and a design note.

Passed all tests. Also checking in modified files, emoji_post_source.xml and font_db.txt.